+++ StreamFiles/StreamFiles/SFManager.m
@@ -16,7 +16,7 @@
 
 @implementation SFManager
 
-+(void) readDataFromPath : (NSString*) path withStream : (const void* (^)(uint8_t* buffer, unsigned int length)) stream completion : (void (^)(BOOL isSuccess, NSData *data)) completion {
++(void) readDataFromPath : (NSString*) path withStream : (const void* (^)(uint8_t* buffer, unsigned long length)) stream completion : (void (^)(BOOL isSuccess, NSData *data)) completion {
     
     SFInput *input = [SFInput new];
     [[self streamPool] addObject:input];
@@ -30,7 +30,7 @@
     
 }
 
-+(void) writeDataToPath : (NSString*) path withData : (NSData*) data withStream : (const uint8_t* (^)(uint8_t* buffer, unsigned int length)) stream completion : (void (^)(BOOL isSuccess)) completion {
++(void) writeDataToPath : (NSString*) path withData : (NSData*) data withStream : (const uint8_t* (^)(uint8_t* buffer, unsigned long length)) stream completion : (void (^)(BOOL isSuccess)) completion {
     
     SFOutput *output = [SFOutput new];
     [[self streamPool] addObject:output];
@@ -46,7 +46,7 @@
 
 +(void) readFromPath : (NSString*) fromPath
          writeToPath : (NSString*) toPath
-          withStream : (const void* (^)(uint8_t* buffer, unsigned int length)) stream
+          withStream : (const void* (^)(uint8_t* buffer, unsigned long length)) stream
           completion : (void (^)(BOOL isSuccess)) completion {
     
     SFOnTheFly *fly = [SFOnTheFly new];
